refactor(manifest): extract ontology validator into parser_ontology.go
Moves validateOntology out of parser.go into its own file to keep
validation adjacent to the Ontology type and prepare for consumption
by the upcoming ontology bounded-context package.
